fix postal & temperature tests
