+++ test/controllers/temperatures_controller_test.rb
@@ -6,31 +6,31 @@ class TemperaturesControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should get index" do
-    get temperatures_url, as: :json
+    get api_v1_temperatures_url, as: :json
     assert_response :success
   end
 
   test "should create temperature" do
     assert_difference("Temperature.count") do
-      post temperatures_url, params: { temperature: { forecast: @temperature.forecast, postal_code_id: @temperature.postal_code_id } }, as: :json
+      post api_v1_temperatures_url, params: { temperature: { forecast: @temperature.forecast, postal_code_id: @temperature.postal_code_id } }, as: :json
     end
 
     assert_response :created
   end
 
   test "should show temperature" do
-    get temperature_url(@temperature), as: :json
+    get api_v1_temperature_url(@temperature), as: :json
     assert_response :success
   end
 
   test "should update temperature" do
-    patch temperature_url(@temperature), params: { temperature: { forecast: @temperature.forecast, postal_code_id: @temperature.postal_code_id } }, as: :json
+    patch api_v1_temperature_url(@temperature), params: { temperature: { forecast: @temperature.forecast, postal_code_id: @temperature.postal_code_id } }, as: :json
     assert_response :success
   end
 
   test "should destroy temperature" do
     assert_difference("Temperature.count", -1) do
-      delete temperature_url(@temperature), as: :json
+      delete api_v1_temperature_url(@temperature), as: :json
     end
 
     assert_response :no_content
